spec/src/requirements.txt
mcp
pyyaml
requests
uvicorn
